Songonku

+++ Minigame/TrainingFramework/src/GameObject/Sprite2D.cpp
@@ -143,9 +143,9 @@ void Sprite2D::MovingMonster(GLfloat xVal)
 	}
 }
 
-void Sprite2D::Set2DPosition(GLint x, GLint y)
+void Sprite2D::Set2DPosition(GLfloat x, GLfloat y)
 {
-	m_position = Vector3((float)x, (float)y, 0.0f);
+	m_position = Vector3(x, y, 0.0f);
 	CalculateWorldMatrix();
 }
